pyproject.toml
[project]
name = "qa-two-playwright-pytest-allure"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13.5"
dependencies = [
    "allure-pytest>=2.15.0",
    "faker>=37.6.0",
    "playwright==1.52.0",
    "pre-commit>=4.3.0",
    "pytest-dotenv>=0.5.2",
    "pytest-playwright>=0.7.0",
    "ruff>=0.14.0",
    "transliterate>=1.10.2",
    "vulture>=2.14",
]

[tool.pytest.ini_options]
addopts = "--color=yes"
log_cli = true
markers = [
    "user_login: scenario user login",
    "user_registration: scenario user registration"
]


[tool.ruff]
exclude = [
    ".git",
    ".pyenv",
    ".pytest_cache",
    ".pytype",
    ".ruff_cache",
    ".venv",
    ".vscode",
    "__pypackages__",
    "venv",
]

line-length = 88
indent-width = 4
target-version = "py313"

[tool.ruff.lint]
preview = true
select = [
    "UP",   # pyupgrade
    "I",    # isort
    "E",    # flake8-errors
    "W",    # flake8-warnings
    "F",    # flake8-pyflakes
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "S",    # flake8-bandit
    "BLE",  # flake8-blind-except
    "T20",  # flake8-print
    "T10",  # flake8-debugger
    "PT",   # flake8-pytest-style
    "RUF",  # Ruff
    "DTZ",  # flake8-datetimez
    "ERA",  # eradicate
    "FURB", # refurb
]
ignore = [
    "C408",  # Ненужный вызов dict, list и т.д.
    "RUF001", # Игнорирование похожих символов с разными ASCII
    "RUF002", # Игнорирование похожих символов с разными ASCII
    "RUF003", # Игнорирование похожих символов с разными ASCII
    "ERA001", # Игнорирование коментариев в коде
    # "S110",  # try-except-pass
]
fixable = [
    "I",
]

[tool.ruff.format]
quote-style = "double" 
indent-style = "space"

[tool.vulture]
exclude = [".venv/"]
ignore_decorators = ["@pytest.fixture"]
ignore_names = ["open_login_page"]
# make_whitelist = true
# min_confidence = 80
# #paths = ["myscript.py", "mydir", "whitelist.py"]
# sort_by_size = true
# verbose = true
